# PERF: rebuild_blknos_and_blklocs