--- a/pandas/_libs/internals.pyx
+++ b/pandas/_libs/internals.pyx
@@ -670,6 +670,39 @@ cdef class BlockManager:
         self._blklocs = None
 
     # -------------------------------------------------------------------
+    # Block Placement
+
+    def _rebuild_blknos_and_blklocs(self) -> None:
+        """
+        Update mgr._blknos / mgr._blklocs.
+        """
+        cdef:
+            intp_t blkno
+            cnp.npy_intp length = self.shape[0]
+            SharedBlock blk
+            BlockPlacement bp
+
+        # equiv: np.empty(length, dtype=np.intp)
+        new_blknos = cnp.PyArray_EMPTY(1, &length, cnp.NPY_INTP, 0)
+        new_blklocs = cnp.PyArray_EMPTY(1, &length, cnp.NPY_INTP, 0)
+        new_blknos.fill(-1)
+        new_blklocs.fill(-1)
+
+        for blkno, blk in enumerate(self.blocks):
+            bp = blk.mgr_locs
+            new_blknos[bp.indexer] = blkno
+            new_blklocs[bp.indexer] = np.arange(len(bp))
+
+        for blkno in new_blknos:
+            # If there are any -1s remaining, this indicates that our mgr_locs
+            #  are invalid.
+            if blkno == -1:
+                raise AssertionError("Gaps in blk ref_locs")
+
+        self._blknos = new_blknos
+        self._blklocs = new_blklocs
+
+    # -------------------------------------------------------------------
     # Pickle
 
     cpdef __reduce__(self):