… pandas/_libs/internals.pyi